requirements.txt
fastapi
uvicorn
jinja2
python-multipart
sqlalchemy
pydantic
python-dotenv
httpx
authlib
itsdangerous

